refactored testrunner code into its own package
