+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/exercism/go-test-runner
 
 go 1.15
+
+require (
+	github.com/gopherdata/gophernotes v0.7.1 // indirect
+	golang.org/x/tools v0.0.0-20210115202250-e0d201561e39 // indirect
+)
