+++ main_test.go
@@ -1,104 +1,97 @@
-package gotestrun
+package main
 
 import (
-	"encoding/json"
 	"fmt"
+	"os"
+	"os/exec"
 	"strings"
 	"testing"
 )
 
-func TestRunTests_broken(t *testing.T) {
-	input_dir := "./testdata/practice/broken"
-	cmdres, ok := runTests(input_dir)
-	if ok {
-		t.Errorf("Broken test did not fail %s", cmdres.String())
+func ExampleMain() {
+	os.Args = []string{"path", "testrunner/testdata/practice/passing", "outdir"}
+	main()
+	// Output:
+}
+
+func TestMain_fail(t *testing.T) {
+	tests := []struct {
+		name    string
+		args    []string
+		exitMsg string
+	}{
+		{
+			name: "incorrect usage",
+			args: []string{"progpath", "only_one_arg"},
+		},
+		{
+			name: "missing input_dir",
+			args: []string{"progpath", "bad_input_dir", "noop"},
+		},
 	}
 
-	res := cmdres.String()
-	pre := "FAIL	github.com/exercism/go-test-runner/testdata/practice/broken [build failed]"
-	post := "returned exit code 2: exit status 2"
-	if !strings.HasPrefix(res, pre) {
-		t.Errorf("Broken test run - unexpected prefix: %s", res)
-	}
-	if !strings.HasSuffix(res, post) {
-		t.Errorf("Broken test run - unexpected suffix: %s", res)
-	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// technique from: https://talks.golang.org/2014/testing.slide#23
+			// base case - if the env var is set, just run main()
+			if os.Getenv("BE_MAIN") == "1" {
+				os.Args = tt.args
+				main()
+				return
+			}
 
-	output := &testReport{
-		Status:  statErr,
-		Message: res,
-	}
-	btr, err := json.MarshalIndent(output, "", "\t")
-	if err != nil {
-		t.Errorf("Broken test output not valid json: %s", err)
-	}
-	tr := string(btr)
-
-	pre = `{
-	"status": "error",
-	"message": "FAIL\tgithub.com/exercism/go-test-runner/testdata/practice/broken [build failed]`
-
-	post = `returned exit code 2: exit status 2",
-	"tests": null
-}`
-	if !strings.HasPrefix(tr, pre) {
-		t.Errorf("Broken test run unexpected json prefix: %s", tr)
-	}
-	if !strings.HasSuffix(tr, post) {
-		t.Errorf("Broken test run unexpected json suffix: %s", tr)
+			// exec the test from itself, so we can safely inspect the fatal exit
+			tn := fmt.Sprintf("-test.run=TestMain_fail/%s", tt.name)
+			cmd := exec.Command(os.Args[0], tn)
+			// set the base case env var to avoid infinite recursion
+			cmd.Env = append(os.Environ(), "BE_MAIN=1")
+			err := cmd.Run()
+			if e, ok := err.(*exec.ExitError); ok || !e.Success() {
+				return
+			}
+			t.Fatalf("main() test ran with err %v, want exit status 1", err)
+		})
 	}
 }
 
-func ExampleRunTests_passing() {
-	input_dir := "./testdata/practice/passing"
-
-	cmdres, ok := runTests(input_dir)
-	if !ok {
-		fmt.Printf("Passing test failed: %s", cmdres.String())
+func TestCheckArgs(t *testing.T) {
+	tests := []struct {
+		name       string
+		input_dir  string
+		output_dir string
+		ok         bool
+		msg        string
+	}{
+		{
+			name:       "missing input_dir",
+			input_dir:  "testrunner",
+			output_dir: "testrunner",
+			ok:         true,
+			msg:        "",
+		},
+		{
+			name:       "missing input_dir",
+			input_dir:  "bad_input_dir",
+			output_dir: "noop",
+			ok:         false,
+			msg:        "input_dir does not exist: bad_input_dir",
+		},
+		{
+			name:       "broken output_dir",
+			input_dir:  "testrunner",
+			output_dir: "/tmp/broken/rmme",
+			ok:         false,
+			msg:        "output_dir /tmp/broken/rmme does not exist, mkdir failed:",
+		},
 	}
-
-	output := getStructure(cmdres, input_dir)
-	if bts, err := json.MarshalIndent(output, "", "\t"); err != nil {
-		fmt.Printf("Passing test output not valid json: %s", err)
-	} else {
-		fmt.Println(string(bts))
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			msg, ok := checkArgs(tt.input_dir, tt.output_dir)
+			if ok != tt.ok || !strings.HasPrefix(msg, tt.msg) {
+				t.Fatalf("checkArgs(%s, %s) = %s, %t, want %s, HasPrefix(%t)",
+					tt.input_dir, tt.output_dir, msg, ok, tt.msg, tt.ok,
+				)
+			}
+		})
 	}
-	// Output: {
-	//	"status": "pass",
-	//	"tests": [
-	//		{
-	//			"name": "TestTrivialPass",
-	//			"status": "pass",
-	//			"test_code": "// Trivial passing test example\nfunc TestTrivialPass(t *testing.T) {\n\tif true != true {\n\t\tt.Fatal(\"This was supposed to be a tautological statement!\")\n\t}\n\tfmt.Println(\"sample passing test output\")\n}",
-	//			"message": "\n=== RUN   TestTrivialPass\n\nsample passing test output\n\n--- PASS: TestTrivialPass (0.00s)\n"
-	//		}
-	//	]
-	//}
-}
-
-func ExampleRunTests_failing() {
-	input_dir := "./testdata/practice/failing"
-
-	cmdres, ok := runTests(input_dir)
-	if !ok {
-		fmt.Printf("Failing test expected to return ok: %s", cmdres.String())
-	}
-
-	output := getStructure(cmdres, input_dir)
-	if bts, err := json.MarshalIndent(output, "", "\t"); err != nil {
-		fmt.Printf("Failing test output not valid json: %s", err)
-	} else {
-		fmt.Println(string(bts))
-	}
-	// Output: {
-	//	"status": "fail",
-	//	"tests": [
-	//		{
-	//			"name": "TestTrivialFail",
-	//			"status": "fail",
-	//			"test_code": "// Trivial failing test example\nfunc TestTrivialFail(t *testing.T) {\n\tif false != true {\n\t\tt.Fatal(\"Intentional test failure\")\n\t}\n\tfmt.Println(\"sample failing test output\")\n}",
-	//			"message": "\n=== RUN   TestTrivialFail\n\n    failing_test.go:11: Intentional test failure\n\n--- FAIL: TestTrivialFail (0.00s)\n"
-	//		}
-	//	]
-	//}
 }
