+++ testrunner/ast_test.go
@@ -1,4 +1,4 @@
-package gotestrun
+package testrunner
 
 import (
 	"testing"
